+++ R/arkhe-deprecated.R
@@ -1,6 +1,14 @@
+# DATA CLEANING: APPEND
+#' @include assert.R
+NULL
+
 #' Deprecated Functions in arkhe
 #'
 #' These functions still work but will be removed (defunct) in the next version.
 #' @name arkhe-deprecated
 #' @keywords internal
 NULL
+
+#' @export
+#' @rdname arkhe-deprecated
+assert_dimensions <- assert_dim